chore: move refresh app to general actions file
