+++ src/actions/vue/general.ts
@@ -1,0 +1,3 @@
+export const refreshApp = (): Promise<string> => new Promise((resolve) => {
+  resolve(window.ipcRenderer.invoke('refresh-app'))
+})
